mod: refactor calls
