+++ backend/packages/vesting-scan/src/scan/constants.js
@@ -1,0 +1,16 @@
+module.exports = {
+  SECTION: {
+    VESTING: "vesting",
+  },
+  EXTRINSIC_METHOD: {
+    VESTED_TRANSFER: "vestedTransfer",
+    FORCE_VESTED_TRANSFER: "forceVestedTransfer",
+    VEST: "vest",
+    VEST_OTHER: "vestOther",
+    MERGE_SCHEDULES: "mergeSchedules",
+  },
+  EVENT_METHOD: {
+    VESTING_UPDATED: "VestingUpdated",
+    VESTING_COMPLETED: "VestingCompleted",
+  },
+};
