+++ backend/packages/vesting-scan/src/scan/extrinsics/calls/index.js
@@ -1,0 +1,13 @@
+const { handleVestedTransfer } = require("./vestedTransfer");
+const { handleForceVestedTransfer } = require("./forceVestedTransfer");
+const { handleVest } = require("./vest");
+const { handleVestOther } = require("./vestOther");
+const { handleMergeSchedules } = require("./mergeSchedules");
+
+module.exports = {
+  handleVestedTransfer,
+  handleForceVestedTransfer,
+  handleVest,
+  handleVestOther,
+  handleMergeSchedules,
+};
